Module Imandra_codegen__Gen_json

Code Generation for Json (de)serializers

This module provides serialization/deserialization for all types defined in logic mode.

A json module should have the module type JSON.

module type JSON = sig ... end

val mk_plugin : ?⁠logic_deser:bool -> mod_name:string -> unit -> Imandra_surface.Plugin.t

Create a plugin for Json (de)ser.

parameter mod_name

the name of the Json module, must satisfy the signature JSON

parameter logic_deser

if true, deserializers are in logic mode and will be validated. Careful: can be slower