Module Imandra_interactive.Decompose

Region Decomposition

type t = private Imandra_surface.Top_result.decompose_region = {

reg_session : Imandra_surface.Top_result.decompose_session_id;

reg_id : int;

reg_args : Imandra_surface.Var.t list;

reg_constraints : Imandra_surface.Term.t list;

reg_invariant : Imandra_surface.Term.t;

reg_liveness : Imandra_surface.Decompose_region.liveness;

}

type 'a gen = unit -> 'a option

type context

A context is a transient handle to the server's decomposition state. It can only be created via top_gen and should NEVER escape the scope of the function argument.

val context_of_region : t -> context

val context_next : context -> t option

Next region in this context, in case it comes from top_incremental.

val context_close : context -> unit

Release the context (and resources, such as memory, associated with it). From now on, get_model cannot be called on any region that depends on this context.

val gen_of_context : context -> t gen

val get_model : t -> Imandra_surface.Top_result.term_model

Get a model for the given region. If the associated context (see context_of_region) has been closed with context_close or because top_incremental has returned, this will raise an error.

val top_incremental : ?⁠compound:bool -> ?⁠max_ctx_simp:int -> ?⁠basis:string list -> ?⁠assuming:string -> ?⁠interpret_basis:bool -> ?⁠ctx_asm_simp:bool -> ?⁠reduce_symmetry:bool -> ?⁠aggressive_rec:bool -> string -> (context -> 'a) -> 'a

top_incremental f callback decomposes function f lazily. It calls callback with a context, that can be used to iterate over the regions of f, obtain models. etc. callback can use the context as pleases it, but the context becomes invalid after callback ctx returns and MUST NOT be used any further. Once the callback has returned, the session is properly disposed of and resources (memory) released.

val top_ctx : ?⁠compound:bool -> ?⁠max_ctx_simp:int -> ?⁠basis:string list -> ?⁠assuming:string -> ?⁠interpret_basis:bool -> ?⁠ctx_asm_simp:bool -> ?⁠reduce_symmetry:bool -> ?⁠aggressive_rec:bool -> string -> context

top_ctx f decomposes function f lazily. The context must be manually iterated over in order to obtain regions of f, obtain models etc. The context MUST be manually closed using context_close in order to release resources and dispose of the session.

val top_gen : ?⁠compound:bool -> ?⁠max_ctx_simp:int -> ?⁠basis:string list -> ?⁠assuming:string -> ?⁠interpret_basis:bool -> ?⁠ctx_asm_simp:bool -> ?⁠reduce_symmetry:bool -> ?⁠aggressive_rec:bool -> string -> context * t gen

top_gen f decomposes function f lazily. The context is automatically closed once the generator reaches its end, if the generator is not fully consumed, the user MUST manually close the context using context_of_region and context_close in order to release resources and dispose of the session.

val top : ?⁠compound:bool -> ?⁠max_ctx_simp:int -> ?⁠basis:string list -> ?⁠assuming:string -> ?⁠interpret_basis:bool -> ?⁠ctx_asm_simp:bool -> ?⁠reduce_symmetry:bool -> ?⁠aggressive_rec:bool -> string -> t list

Decompose the function whose name is given into regions.

parameter basis

blocked functions

parameter compound

do not decompose compound path conditions

parameter interpret_basis

give definitional interpretations to all basis functions during feasibility checks and sample point extraction

parameter assuming

function (with compatible signature) to assume as a side-condition

parameter reduce_symmetry

apply aggressive symmetry reduction during simplification

parameter aggressive_rec

aggressively manipulates recursive functions appearing in decomposition targets

parameter max_ctx_simp

maximal number of iterations of contextual simplification (pre-lifting)

parameter ctx_asm_simp

apply contextual simplification to the assuming function

parameter max_depth

max depth for the analysis

val string_of_region : t -> string

val print : Format.formatter -> t -> unit

val print_opaque : Format.formatter -> t -> unit

val refine : t -> basis:string list -> Imandra_surface.Term.t list -> t option