Module Imandra_interactive.Genpp

Code Generation for Pretty Printers

type t

State containing a set of printers for some history snapshot

val make : ?⁠hist:History.State.t -> unit -> t

val to_string : t -> string

val to_file : ?⁠filename:string -> t -> unit

Extract generator code into a valid OCaml file.

parameter filename

the file name (default "pp.ml")

val eval : ?⁠quiet:bool -> ?⁠st:t -> unit -> unit

eval st is the same as System.eval (to_string st). After calling eval st, a module Pp should be accessible in the current Imandra environment.

parameter quiet

if true, evaluation is silent (default false)

parameter force

if true, force re-evaluation even if nothing changed (default false)

val printer_of_type_expr : ?⁠qualified:bool -> t -> Imandra_surface.Type.t -> string

Generate code for a printer of type <this type> -> string

parameter qualified

if true, qualifies the printer with the `Pp.` module

Builtin Printers