Module Imandra_interactive.Help

Interactive Help System

The usual way to access this is through the "help" directive, as so:

# #help ;;
(* displays general help message *)

# #help "some topic";;
(* displays help message for topic *)

val help_system : ?rst:string -> unit -> unit

Display help message on stdout.

parameter rst
if provided, rst is the topic to seek help about

val find_help : string -> (Imandra_document.Document.t * Imandra_document.Document.t) option

Find help for the given topic, as a pair (short,long) of short and long documents

module Register : sig ... end