Module Imandra.Gist

val use : ?timeout:int -> string -> unit

Download a github gist given its unique hash, and #use its content. This is an easy way of sharing small snippets of code.

parameter timeout
max time for a single http request, in seconds

Example:

# Imandra.Gist.use "c-cube/2f9112d566c16ca747dc997165ab5017";;
val f : (Z.t * string) list -> bool = <fun>

This can also be used via #use_gist "hash":

# #use_gist "c-cube/2f9112d566c16ca747dc997165ab5017";;
val f : (Z.t * string) list -> bool = <fun>

val mod_use : ?timeout:int -> name:string -> string -> unit

Download a github gist given its unique hash, and #use its content. This is an easy way of sharing small snippets of code.

parameter timeout
max time for a single http request, in seconds

parameter name
a valid OCaml module name

Example:

# Imandra.Gist.mod_use ~name:"Foo" "c-cube/2f9112d566c16ca747dc997165ab5017";;
module Foo : sig val f : (Z.t * string) list -> bool end