Module Imandra_interactive.Instance

Verification of properties

type t = Imandra_surface.Top_result.instance_result =

| I_unsat of {

proof : Imandra_surface.Document.t option;

proof_graph : string lazy_t option;

callgraph : Imandra_surface.Top_result.callgraph option;

}

| I_unsat_upto of {

callgraph : Imandra_surface.Top_result.callgraph option;

upto : Imandra_surface.Event.upto;

}

| I_sat of {

model_str : Imandra_surface.Top_result.model_string;

model : Imandra_surface.Top_result.term_model;

callgraph : Imandra_surface.Top_result.callgraph option;

proof : Imandra_surface.Document.t option;

}

| I_unknown of {

callgraph : Imandra_surface.Top_result.callgraph option;

instances : Imandra_surface.Document.t option;

proof : Imandra_surface.Document.t option;

reason : string;

}

Result of a verification call

val top : ?⁠full:bool -> ?⁠fc:bool -> ?⁠upto:Imandra_surface.Event.upto -> ?⁠quiet:bool -> ?⁠simp:bool -> ?⁠hints:Imandra_surface.Uid.t Imandra_surface.Hints.t -> ?⁠reflect:bool -> string -> t

Main entry point for finding an instance satisfying a property (a boolean function)

parameter reflect

if true, the instance is reflected in a "CX" module.

parameter simp

if true, simplify goal before starting the proof search

parameter fc

if true, use forward chaining rules

parameter full

if true, proofs/callgraphs are computed (at a performance cost). Default false.

For example, here we define a property named "f" (which turns out to have an instance), and verify it.

  (* satisfiable property *)
# let f x = x + 1 = 100 ;;
val f : Z.t -> bool = <fun>

# Instance.top ~reflect:true "f" ;;
Counterexample:
 let x = 99
module CX : sig val x : Z.t end
- : Instance.t = I_sat …

# CX.x ;;
- : Z.t = 99

(* it's an instance alright! *)
# f CX.x ;;
- : bool = true

We obtain a counter-example (as a I_sat value) containing a model), but Imandra also reflects the instance in the OCaml environment as a CX module with a value x : Z.t. This instance can be used to compute, like any other OCaml value.