Module Imandra_interactive.Undo

type snap = Interactive_snapshot.Gen_snap.t

val save_snap0 : unit -> unit

val with_save : (commit:(unit -> unit) -> 'a) -> 'a

with_save f calls f ~commit and returns the same result as this call. It creates a speculative undo point before calling f; if f calls commit() then this undo point is saved, otherwise it is discarded.

exception E_undo of {

count : int;

}

val pop_n : int -> snap option

Pop n snapshots

val undo_by_count : int -> unit

val undo_parse_count : string -> unit

val undo_back_through : string -> unit

undo back through

module Effect : sig ... end