Module Imandra_interactive.Undo_snap

type t

val pp : Ppx_deriving_runtime.Format.formatter -> t -> Ppx_deriving_runtime.unit

val show : t -> Ppx_deriving_runtime.string

val latest : unit -> t

Latest snapshot

include Imandra_util.Snapshot_sigs.S

val push : unit -> unit

val pop_restore : unit -> unit

val pop_delete : unit -> unit

val is_empty : unit -> bool

val with_push : restore_if_ok:bool -> (unit -> 'a) -> 'a

val save_named : int -> unit

val restore_named : int -> unit

val delete_named : int -> unit

val pp_named_debug : int CCFormat.printer

exception E_undo of int

exception used to escape some local snapshot scopes. Applying undo from within these scopes would be no use, so we escape with an exception and undo from outside.