Module Imandra_prelude.Multiset

Multiset

A multiset is a collection of elements that don't have any particular order, but can occur several times (unlike a regular set).

type +'a t = ( 'a, int ) Map.t

val empty : 'a t

val add : 'a -> 'a t -> 'a t

val remove : 'a -> 'a t -> 'a t

val mem : 'a -> 'a t -> bool

val find : 'a -> 'a t -> int

val of_list : 'a list -> 'a t