Module Imandra_surface__Ciml

Ciml files

A ciml file is an on-disk certificate for a file. It can be loaded to accelerate re-checking the file.

module CCh = Imandra_surface.Check_cache

type t

type ciml = t

val serde : t Imandra_util.Serialize.t

val ciml_of_iml : string -> string

Path of the ciml for the given iml file

val read_file : string -> (t, string) Pervasives.result

Open the given file and get a ciml object from it

val write_file : t -> string -> unit

Write into given file

val create : string -> t

val add_entry : t -> CCh.key -> CCh.entry -> unit

val of_entries : string -> CCh.pair list -> t

Build from filename and entries

val entries : t -> (CCh.key * CCh.entry) list

val mtime : t -> float

Last modification time of the file this is a cache of

val length : t -> int

module Cache : sig ... end