Module Imandra_surface__Count_funs

Count functions, for types that need it

val mk_count_fun_ty : ?⁠db:Imandra_surface.Event.DB.t -> ?⁠find_cnt:(Imandra_surface.Uid.t -> Imandra_surface.Event.fun_data option) -> Imandra_surface.Type.t -> Imandra_surface.Event.fun_data

takes ty and builds a function ty -> int

val mk_count_fun_ty_as_lam : ?⁠db:Imandra_surface.Event.DB.t -> ?⁠find_cnt:(Imandra_surface.Uid.t -> Imandra_surface.Event.fun_data option) -> Imandra_surface.Type.t -> Imandra_surface.Term.t

Same as mk_count_fun_ty but returns a lambda term.

val mk_count_fun_tuple : ?⁠db:Imandra_surface.Event.DB.t -> ?⁠find_cnt:(Imandra_surface.Uid.t -> Imandra_surface.Event.fun_data option) -> Imandra_surface.Type.t list -> Imandra_surface.Event.fun_data

takes ty1…tyn and builds a function (ty1 * … * tyn) -> int

val mk_count_funs : ?⁠db:Imandra_surface.Event.DB.t -> ?⁠find_cnt:(Imandra_surface.Uid.t -> Imandra_surface.Event.fun_data option) -> Imandra_surface.Type.def list -> (Imandra_surface.Uid.t * Imandra_surface.Term.t * Imandra_surface.Event.fun_data) list

takes type definitions and returns the type name, a function term and function definition for each of the type def.