Module Imandra_surface__Valgen

Randomized type inhabitant generation for property testing for ImandraML.

type 'a t = Random.State.t -> 'a

val bool : bool t

val int : ?⁠n:int -> int t

val float : ?⁠f:float -> float t

val alpha : char t

val string : ?⁠len:int -> string t

val list : ?⁠len:int t -> 'a t -> 'a list t

val opt : 'a t -> 'a option t

Properties

module Prop : sig ... end

Properties to be tested on a number of random test cases

val default_seed : int array

type 'a result =

| Passed of int

| Failed of 'a list

| Error of 'a option * exn

val check : ?⁠prng:Random.State.t -> ?⁠n:int -> 'a t -> 'a Prop.t -> 'a result