Module Rewrite_parsetree.Custom_extension

Custom extensions defined by users

module U : sig ... end

type expr_tr = self:(Parsetree.expression -> Parsetree.expression) -> Parsetree.expression -> Parsetree.expression

type value_binding_tr = self_expr:(Parsetree.expression -> Parsetree.expression) -> (string * (string * Parsetree.core_type option) list * Parsetree.expression) -> Parsetree.value_binding list

type string_lit_tr = self_expr:(Parsetree.expression -> Parsetree.expression) -> string -> Parsetree.expression

Transformation on string literals like {foo|…|foo}

type tr =

| Tr_e of expr_tr
transformation for extension points %foo …

| Tr_vb of value_binding_tr
transformation for value bindings let%foo …

| Tr_string_lit of string_lit_tr
transformation for string literals {foo|…|foo}

| Tr_id

val add_expr_tr : string -> expr_tr -> unit

Custom_extension.add_expr_tr name f register f as a function to be called on e to transform [%name e] into some other expression f e. This is useful for writing custom blocks.

val add_vb_tr : string -> value_binding_tr -> unit

val add_string_lit_tr : string -> string_lit_tr -> unit

add_string_lit_tr "foo" f calls f on literals {foo|…|foo}

val with_local : (string * tr) list -> (unit -> 'a) -> 'a

Locally define the given list of extensions, call f(), remove extensions, and return f()'s result.