Module type Fut.S0

Basic operations on futures

type 'a t

The type of futures values of type 'a

val pp : 'a CCFormat.printer -> 'a t CCFormat.printer

val state : 'a t -> 'a fut_state

val is_done : _ t -> bool

is_done fut is true iff state fut is not Fut_wait. In other words, the future has resolved to a value or an exception.

val return : 'a -> 'a t

Future that just contains a value

val fail : exn -> 'a t

Future that just contains a failure

val fail_with_backtrace : exn -> backtrace -> 'a t

Future that just contains a failure and backtrace

val fail_with_bt_opt : exn -> backtrace option -> 'a t

val error : ?⁠process:string -> ?⁠kind:Imandra_util.Error.kind -> ?⁠loc:Imandra_util.Iloc.t -> string -> _ t

Interface to Error.error

val errorf : ?⁠process:string -> ?⁠kind:Imandra_util.Error.kind -> ?⁠loc:Imandra_util.Iloc.t -> ('a, Stdlib.Format.formatter, unit, 'b t) Stdlib.format4 -> 'a

Interface to Error.error

val return_res : ('a, exn_with_bt_opt) Stdlib.result -> 'a t

val wait_block : 'a t -> 'a

Block the current thread until the future is resolved.

raises an

exception e if the future was resolved with set_exn or created with fail

val wait_block_catch : 'a t -> ('a, exn_with_bt_opt) Stdlib.result

Block the current thread until the future is resolved, like wait. Catches exceptions into a result.

val wait_block_poll : 'a t -> 'a

Wait in an active poll, so as to still catch signals and check other things.

val ignore : _ t -> unit