Module Imandra_thread_util__Background_thread

Background thread, for running tasks asynchronously

type t

A background thread

val start : name:string -> unit -> t

val call : t -> ('a -> 'b) -> 'a -> 'b Imandra_thread_util.Fut.t

call thread f x queues a task that will compute f x, and returns a future for the result of that computation.

val call0 : t -> (unit -> 'a) -> 'a Imandra_thread_util.Fut.t

val call1 : t -> ('a -> 'b) -> 'a -> 'b Imandra_thread_util.Fut.t

val call2 : t -> ('a -> 'b -> 'c) -> 'a -> 'b -> 'c Imandra_thread_util.Fut.t

val call_ignore : t -> ('a -> unit) -> 'a -> unit

call_ignore thread f x queues a task that will call f x, and discards the result. NOTE exceptions raised by the task are ignored for now.

val call0_ignore : t -> (unit -> unit) -> unit

val call1_ignore : t -> ('a -> unit) -> 'a -> unit

val call2_ignore : t -> ('a -> 'b -> unit) -> 'a -> 'b -> unit