Module type Imandra_thread_util__TPool.S

Signature of the thread pool

include Imandra_thread_util.Fut.EXECUTOR

val run : (unit -> unit) -> unit

run f schedules f to run on some background thread

val run1 : ('a -> unit) -> 'a -> unit

run1 f x schedules f x to run on some background thread

val run2 : ('a -> 'b -> unit) -> 'a -> 'b -> unit

val set_exn_handler : (exn -> Stdlib.Printexc.raw_backtrace -> unit) -> unit

module Fut : Imandra_thread_util.Fut.S