Module Imandra_util.Profile

type probe

val create : name:string -> probe

val enter : probe -> unit

val exit : probe -> unit

val with_ : probe -> (unit -> 'a) -> 'a

val with1 : probe -> ('a -> 'b) -> 'a -> 'b

val with2 : probe -> ('a -> 'b -> 'c) -> 'a -> 'b -> 'c

val with3 : probe -> ('a -> 'b -> 'c -> 'd) -> 'a -> 'b -> 'c -> 'd

val name : probe -> string

val get_all : unit -> probe list

All registered probes

val total_time : probe -> float

Total processor time spent in this probe, in seconds

val total_alloc : probe -> float

Total number of allocated bytes in this probe

val count : probe -> int

Number of times enter was called on this probe

val wall_clock : unit -> float