Module Imandra_util__IO_chunk.FD

val write_l : Unix.file_descr -> ?⁠pool:Pool.t -> t list -> t list

Write all that can be written without blocking. Returns remaining chunks, possibly modified.

parameter pool

if provided, emptied chunks will be recycled in there.

val read_at_most_once : ?⁠pool:Pool.t -> Unix.file_descr -> t -> int -> int

read_at_most_once ic ch n reads at most n bytes into ch. Returns 0 if reading would block.