Module Help.Register

val top : ?⁠listable:bool -> string -> ?⁠short:Imandra_surface.Document.t -> Imandra_surface.Document.t -> unit

Register help for a new topic

parameter listable

if true, the entry will show up in #help;;, otherwise it's only available through #help <topic>;;

parameter short

if provided, give a short document used in #help;;

val reset : unit -> unit

Remove all help topics.