Module Imandra_interactive.History

type iml =

| Iml_ty of Imandra_surface.Type.def list

| Iml_term of Imandra_surface.Term.t

| Iml_alias of Imandra_surface.Uid.t

| Iml_struct of Imandra_surface.Event.t list

| Iml_import of string * iml list

| Iml_none

type event = Imandra_surface.Event.t

type error = {

loc : Imandra_util.Iloc.t option;

msg : string;

}

module State : sig ... end

val record_event : ?⁠do_snap:bool -> event -> unit

val record_event_l : ?⁠do_snap:bool -> event list -> unit

val record_error : ?⁠loc:Imandra_util.Iloc.t -> string -> unit

val record_exn : unit -> unit

val record_imported_file : string -> unit

val file_imported : string -> bool

val files_imported : unit -> Imandra_util.Util.Str_set.t

val exn_seen : unit -> bool

val n_exn_seen : unit -> int

val error_seen : unit -> bool

val n_error_seen : unit -> int

val show_history : ?⁠ignore_prelude:bool -> ?⁠key:string option -> unit -> unit

Show history on stdout.

parameter key

the particular event to display if Pconfig.State.state.top_results is true, it pushes a top result instead of printing

val show_all_thms : unit -> unit

val print_dot : ?⁠ignore_prelude:bool -> unit CCFormat.printer

Print the history as a dependency graph

val to_string_dot : ?⁠ignore_prelude:bool -> unit -> string

val event_to_doc : ?⁠txt:bool -> event -> Imandra_surface.Document.t

val to_list : ?⁠ignore_prelude:bool -> unit -> event list

val on_zero_landmark_ptr : (unit -> unit) Stdlib.ref

val set_zero_landmark : unit -> unit

val has_seen_zero_landmark : unit -> bool

val completion_source : Imandra_surface.Completion.source

completion using logic-mode defs

val num_to_ubt_for_name : string -> int

val iter_plugins : ?⁠ignore_prelude:bool -> ?⁠hist:State.t -> acts:Imandra_surface.Plugin.actions -> Imandra_surface.Plugin.set -> unit

Apply the plugins to each even in hist with actions acts.