Module Imandra_interactive.Interactive_snapshot

Snapshot of all of Interactive

This module controls the state of most of Imandra_interactive. It allows push/pop control of the state, for transactions, and named snapshots for #undo directives or other forms of explicit state checkout.

Snapshot for Globals_.event_state_

include Imandra_util.Snapshot_sigs.S

val push : unit -> unit

val pop_restore : unit -> unit

val pop_delete : unit -> unit

val is_empty : unit -> bool

val with_push : restore_if_ok:bool -> (unit -> 'a) -> 'a

val save_named : int -> unit

val restore_named : int -> unit

val delete_named : int -> unit

module Gen_snap : sig ... end

Generation of snapshots with unique names