Module Imandra_interactive.Tlcontext

Toplevel globals and low-level contextual book-keeping

type exec_level =

| INTERACT
toplevel / interactive interpreter

| NO_INTERP
no interpreter, e.g., for -d(ependency)

Mode of execution (interactive or not)

val tick : unit -> int

number of phrases interpreted

val exec_level : exec_level Stdlib.ref

val update_exec_level : exec_level -> unit

val print_type : Types.type_expr CCFormat.printer

Pretty print a type

val string_of_ml_type : Types.type_expr -> string

Pretty print a type into a string

val term_of_string : string -> Imandra_surface.Term.t

Parse, type, and scope a string into an Imandra term

val syn_term_of_string : string -> Imandra_surface.Term.Syn.t

Parse, type, and scope a string into an Imandra syntactic term

val quit : unit -> 'a

Quit execution

Execution

val execute_phrase : ?⁠quiet:bool -> ?⁠out:Stdlib.Format.formatter -> Parsetree.toplevel_phrase -> bool

execute_phrase p executes the given parsetree phrase (directive or OCaml structure) and returns true upon success, false upon failure.

parameter quiet

if true, execution is silent; if false, results are printed

val execute_typed_phrase : ?⁠quiet:bool -> ?⁠out:Stdlib.Format.formatter -> old_env:Env.t -> new_env:Env.t -> Tt.typed_top_phrase -> bool

Execute the given typedtree structure.

parameter quiet

if true, execution is silent; if false, results are printed

module Dirs_printer : sig ... end

val eval_str : ?⁠quiet:bool -> string -> unit

Parse and evaluate a string containing Imandra code.

parameter quiet

if true, execution is silent; if false, results are printed

val require_lib : ?⁠meth:[ `Mod_use | `Import | `Use ] -> ?⁠syntax:Imandra_syntax.Syntax.t -> string -> unit

Require an OCaml library, and possibly also load its imandra file using the given syntax and method of loading.

module Gist : sig ... end

val cur_prompt : unit -> string

Current prompt

val install_doc : string -> unit

Install given function as a document printer

val remove_doc : string -> unit

Removes given function as a document printer

val code_to_exec : Imandra_surface.Code_fragment.t Stdlib.Queue.t

Queue of code fragments to execute

Printing

val print_timer : unit -> unit

Print last timer on stdout

val print_directives : unit -> unit

Print available toplevel directives on stdout